Fix insert with cur_size == max_size
Fixed a bug that occurred when inserting with cur_size == max_size. See
#3
Note this fix is not optimal wrt. memory usage, as it may increase the
inner LRU's capacity although it is not strictly necessary.

I also introduced the test case from the issue description to check for
any re-introduction of the error in the future.